requirements.txt
qdrant-client
python-dotenv
sentence-transformers
transformers
llama-index-core
llama-index-embeddings-huggingface
